hostnamectl: ignore a dot at the end of valid hostnames
When setting the hostname using FQDN including the trailing dot (for
the root zone), which can be omitted as stated in RFC 1034 section 3.1,
set-hostname would interpret the hostname differently compared to when
the trailing dot is omitted. If the pretty hostname is a valid hostname,
simply ignore the dot at the end if it is there.

https://bugzilla.redhat.com/show_bug.cgi?id=1238246