path-util: Add path_normalize() function
The name "normalize" was chosen to be analogous to python's
`os.path.normpath()`.

This modifies the path in place to strip '.' components and collapse
'..' components.

It happens to also collapse multiple / separators into one, because it
was more robust to do so than assert the input is only single /
separated, and when normalized it should only have one /.

This is necessary because glibc's canonical_file_name requires the
target to exist, and for some circumstances we need a normalized path to
a target that does not yet exist.