bitmap: fix bitmap_set on bitmaps which have been cleared
Given two bitmaps and the following code:

  Bitmap *a = bitmap_new(), *b = bitmap_new();
  bitmap_set(a, 1);
  bitmap_clear(a);
  bitmap_set(a, 0);
  bitmap_set(b, 0);

These two bitmaps should now have the same bits set and they should be
equal but bitmap_equal() will return false in this case because while
bitmap_clear() resets the number of elements in the array it does not
clear the array and bitmap_set() expects the array to be cleared.
GREEDY_REALLOC0 looks at the allocated size and not the actual size so
it does not clear any memory.

Fix this by doing an explicit memzero() in bitmap_set() when n_bitmap
gets increased, also since we now always do memzero() ourself anyway we
can change from GREEDY_REALLOC0 to GREEDY_REALLOC.

This also adds test code for this issue.