rfkill: Rework systemd-rfkill
This commit adds a udev rule to save the RFKill state on every change,
so systemd-rfkill always have the most up-to-date state. This also
removes the need for saving the RFKill state on shutdown, so
systemd-rfkill@.service does not have to be active until shutdown, which
in turn removes the need for the RFKill device units.

Without this change systemd-rfkill does not support devices that lose
power over suspend and do a disconnect()/probe() cycle (when the driver
does not implement a reset_resume() callback): systemd-rfkill will
restore the RFKill state that was saved on the last shutdown instead of
the one right before suspending.

This also removes an item from our TODO.