nspawn: fix using --template with --machine
This works:
$ /usr/bin/systemd-nspawn -b --template=/var/lib/machines/f23 --directory=/var/lib/machines/test

However, this doesn't:
$ /usr/bin/systemd-nspawn -b --template=/var/lib/machines/f23 --machine=test
No image for machine 'test': No such file or directory

They should end in the same result; a snapshot (or copy) of "f23" should
be created as "test".

This commit modifies nspawn so that they are handled in the same way
under the covers.

Fixes #1018