cg_migrate_everywhere: set ignore_self to true
Pid 1 is migrating itself when doing cg_migrate_everywhere.  This
particularly becomes a problem when running under cgmanager, where
it is no longer allowed to return itself to the parent cgroup later.
The result is that, for instance, after starting cgproxy inside a
systemd-based container, /proc/1/cgroup shows:

12:name=systemd:/user/1000.user/1.session/lxc/w3
5:devices:/user/1000.user/1.session/lxc/w3/system.slice/cgproxy.service

This was introduced by commit 4ad4900, which iiuc changed the behavior
from ignoring self to not ignoring self.  I believe this was an
inadvertent change, so propose un-doing it.

With this patch, everything appears to function as normal, and pid 1
stays in

12:name=systemd:/user/1000.user/1.session/lxc/w3
5:devices:/user/1000.user/1.session/lxc/w3

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>