core: Replace FOREACH_WORD_QUOTED with strv_split_extract in config_parse_cpu_affinity2
This addresses the TODO to get rid of FOREACH_WORD_QUOTED.

Tested by setting `CPUAfinity=1 2' in /etc/systemd/system.conf, booting
the system with it and checking that /proc/1/status indicates only CPUs
1 and 2 are allowed for PID 1.

No regressions observed in test cases.

Note: Detection of trailing garbage was lost, since strv_split_extract
does not advance the rvalue pointer, so it does not make it possible to
detect whether some part of it was ignored. Perhaps this should be
addressed by updating strv_split_extract to return that information?