test: udev - Skip the test if Perl is not installed.
Some distributions (like Fedora) no longer install Perl by default.

The general build of systemd will require Perl (for automake, intltool,
etc.) but it is possible to build from a dist archive, with particular
configuration options (e.g. --disable-nls, etc.) so that Perl is not
required in any of the steps.

It is still required to run this single udev-test.pl. Fix that by making
the test detect whether Perl is available and skipping the test (with an
`exit 77` command) in case it is not available.

The recipe is inspired by similar ones in the `perlrun` manpage,
particularly in the top of the description (suggesting the use of -x to
ignore the initial shebang) and on the -S option (indicating the script
should be looked for in $PATH in case $0 is incomplete.)

Tested by simulating Perl not being installed (chmod 0 /usr/bin/perl)
and running both `make check` and the test/udev-test.pl script manually.